DiscordListBot/requirements.txt
aiohttp==0.21.6
astroid==1.4.8
cffi==1.8.3
chardet==2.3.0
colorama==0.3.7
discord.py==0.11.0
future==0.15.2
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.5.2
oauthlib==1.0.3
pycparser==2.14
pylint==1.6.4
PyNaCl==1.0.1
python-twitter==3.0rc1
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
websockets==3.1
wrapt==1.10.8
